requirements.txt
backports.tarfile
black
fqdn
gymnasium
importlib-metadata
isoduration
isort
jaraco-collections
jaraco-functools
jaraco.collections
jsonpointer
jupyterlab
jupyterlab-code-formatter
mplfinance
numpy-typing-compat
openpyxl
overrides
pandas-stubs
pip-chill
pip-review
pyside6
rfc3987-syntax
scipy
scipy-stubs
seaborn
tinycss2
tomli
torch
uri-template
webcolors
xlrd
xlwings
xlwt
yfinance
